Create product-of-diff
Takes a list of integers (a_1 a_2 ... a_n) and computes the product of all terms:
(a_i - a_j), where i < j that is, computes the product of all differences between elements, with the element appearing later on the list subtracted from the element appearing first. For example for the list:
(3 2 1)
the function computes:
(3-1)*(3-2)*(2-1) = 2
If the list is empty, the function returns 1.
Name the function product-of-diff.